renamed abci/core.py and class names, merged utils files
