+++ planetmint/commands/planetmint.py
@@ -25,13 +25,12 @@ from transactions.common.transaction import Transaction
 
 
 from planetmint.abci.rpc import ABCI_RPC
-from planetmint.abci.utils import load_node_key
+from planetmint.abci.utils import load_node_key, public_key_from_base64
 from planetmint.application.validator import Validator
 from planetmint.backend import schema
 from planetmint.commands import utils
 from planetmint.commands.utils import configure_planetmint, input_on_stderr
 from planetmint.config_utils import setup_logging
-from planetmint.abci.tendermint_utils import public_key_from_base64
 from planetmint.abci.rpc import MODE_COMMIT, MODE_LIST
 from planetmint.commands.election_types import elections
 from planetmint.version import __tm_supported_versions__
