+++ tests/tendermint/test_core.py
@@ -15,12 +15,11 @@ from transactions import ValidatorElection, ChainMigrationElection
 from transactions.common.crypto import generate_key_pair
 from transactions.types.assets.create import Create
 from transactions.types.assets.transfer import Transfer
-from planetmint.abci.core import App
+from planetmint.abci.application_logic import ApplicationLogic
 from planetmint.backend import query
-from planetmint.abci.core import OkCode, CodeTypeError
+from planetmint.abci.application_logic import OkCode, CodeTypeError
 from planetmint.abci.block import Block
-from planetmint.abci.tendermint_utils import new_validator_set
-from planetmint.abci.tendermint_utils import public_key_to_base64
+from planetmint.abci.utils import new_validator_set, public_key_to_base64
 from planetmint.version import __tm_supported_versions__
 from tests.utils import generate_election, generate_validators
 
@@ -49,7 +48,7 @@ def generate_init_chain_request(chain_id, vals=None):
 
 def test_init_chain_successfully_registers_chain(b):
     request = generate_init_chain_request("chain-XYZ")
-    res = App(b).init_chain(request)
+    res = ApplicationLogic(b).init_chain(request)
     assert res == types.ResponseInitChain()
     chain = query.get_latest_abci_chain(b.models.connection)
     assert chain == {"height": 0, "chain_id": "chain-XYZ", "is_synced": True}
@@ -63,7 +62,7 @@ def test_init_chain_successfully_registers_chain(b):
 def test_init_chain_ignores_invalid_init_chain_requests(b):
     validators = [generate_validator()]
     request = generate_init_chain_request("chain-XYZ", validators)
-    res = App(b).init_chain(request)
+    res = ApplicationLogic(b).init_chain(request)
     assert res == types.ResponseInitChain()
 
     validator_set = query.get_validator_set(b.models.connection)
@@ -77,7 +76,7 @@ def test_init_chain_ignores_invalid_init_chain_requests(b):
     ]
     for r in invalid_requests:
         with pytest.raises(SystemExit):
-            App(b).init_chain(r)
+            ApplicationLogic(b).init_chain(r)
         # assert nothing changed - neither validator set, nor chain ID
         new_validator_set = query.get_validator_set(b.models.connection)
         assert new_validator_set == validator_set
@@ -93,7 +92,7 @@ def test_init_chain_ignores_invalid_init_chain_requests(b):
 def test_init_chain_recognizes_new_chain_after_migration(b):
     validators = [generate_validator()]
     request = generate_init_chain_request("chain-XYZ", validators)
-    res = App(b).init_chain(request)
+    res = ApplicationLogic(b).init_chain(request)
     assert res == types.ResponseInitChain()
 
     validator_set = query.get_validator_set(b.models.connection)["validators"]
@@ -111,7 +110,7 @@ def test_init_chain_recognizes_new_chain_after_migration(b):
     ]
     for r in invalid_requests:
         with pytest.raises(SystemExit):
-            App(b).init_chain(r)
+            ApplicationLogic(b).init_chain(r)
         assert query.get_latest_abci_chain(b.models.connection) == {
             "chain_id": "chain-XYZ-migrated-at-height-1",
             "is_synced": False,
@@ -123,7 +122,7 @@ def test_init_chain_recognizes_new_chain_after_migration(b):
     # a request with the matching chain ID and matching validator set
     # completes the migration
     request = generate_init_chain_request("chain-XYZ-migrated-at-height-1", validators)
-    res = App(b).init_chain(request)
+    res = ApplicationLogic(b).init_chain(request)
     assert res == types.ResponseInitChain()
     assert query.get_latest_abci_chain(b.models.connection) == {
         "chain_id": "chain-XYZ-migrated-at-height-1",
@@ -144,7 +143,7 @@ def test_init_chain_recognizes_new_chain_after_migration(b):
     ]
     for r in invalid_requests:
         with pytest.raises(SystemExit):
-            App(b).init_chain(r)
+            ApplicationLogic(b).init_chain(r)
         assert query.get_latest_abci_chain(b.models.connection) == {
             "chain_id": "chain-XYZ-migrated-at-height-1",
             "is_synced": True,
@@ -161,7 +160,7 @@ def test_init_chain_recognizes_new_chain_after_migration(b):
 
 def test_info(b):
     r = types.RequestInfo(version=__tm_supported_versions__[0])
-    app = App(b)
+    app = ApplicationLogic(b)
 
     res = app.info(r)
     assert res.last_block_height == 0
@@ -174,7 +173,7 @@ def test_info(b):
 
     # simulate a migration and assert the height is shifted
     b.models.store_abci_chain(2, "chain-XYZ")
-    app = App(b)
+    app = ApplicationLogic(b)
     b.models.store_block(Block(app_hash="2", height=2, transactions=[])._asdict())
     res = app.info(r)
     assert res.last_block_height == 0
@@ -187,7 +186,7 @@ def test_info(b):
 
     # it's always the latest migration that is taken into account
     b.models.store_abci_chain(4, "chain-XYZ-new")
-    app = App(b)
+    app = ApplicationLogic(b)
     b.models.store_block(Block(app_hash="4", height=4, transactions=[])._asdict())
     res = app.info(r)
     assert res.last_block_height == 0
@@ -200,7 +199,7 @@ def test_check_tx__signed_create_is_ok(b):
 
     tx = Create.generate([alice.public_key], [([bob.public_key], 1)]).sign([alice.private_key])
 
-    app = App(b)
+    app = ApplicationLogic(b)
     result = app.check_tx(encode_tx_to_bytes(tx))
     assert result.code == OkCode
 
@@ -211,7 +210,7 @@ def test_check_tx__unsigned_create_is_error(b):
 
     tx = Create.generate([alice.public_key], [([bob.public_key], 1)])
 
-    app = App(b)
+    app = ApplicationLogic(b)
     result = app.check_tx(encode_tx_to_bytes(tx))
     assert result.code == CodeTypeError
 
@@ -223,7 +222,7 @@ def test_deliver_tx__valid_create_updates_db_and_emits_event(b, init_chain_reque
 
     tx = Create.generate([alice.public_key], [([bob.public_key], 1)]).sign([alice.private_key])
 
-    app = App(b, events)
+    app = ApplicationLogic(b, events)
 
     app.init_chain(init_chain_request)
 
@@ -253,7 +252,7 @@ def test_deliver_tx__double_spend_fails(b, init_chain_request):
 
     tx = Create.generate([alice.public_key], [([bob.public_key], 1)]).sign([alice.private_key])
 
-    app = App(b)
+    app = ApplicationLogic(b)
     app.init_chain(init_chain_request)
 
     begin_block = types.RequestBeginBlock()
@@ -270,7 +269,7 @@ def test_deliver_tx__double_spend_fails(b, init_chain_request):
 
 
 def test_deliver_transfer_tx__double_spend_fails(b, init_chain_request):
-    app = App(b)
+    app = ApplicationLogic(b)
     app.init_chain(init_chain_request)
 
     begin_block = types.RequestBeginBlock()
@@ -303,7 +302,7 @@ def test_deliver_transfer_tx__double_spend_fails(b, init_chain_request):
 
 
 def test_end_block_return_validator_updates(b, init_chain_request):
-    app = App(b)
+    app = ApplicationLogic(b)
     app.init_chain(init_chain_request)
 
     begin_block = types.RequestBeginBlock()
@@ -335,7 +334,7 @@ def test_end_block_return_validator_updates(b, init_chain_request):
 
 
 def test_store_pre_commit_state_in_end_block(b, alice, init_chain_request):
-    from planetmint.abci.core import App
+    from planetmint.abci.application_logic import ApplicationLogic
     from planetmint.backend import query
 
     tx = Create.generate(
@@ -344,7 +343,7 @@ def test_store_pre_commit_state_in_end_block(b, alice, init_chain_request):
         assets=[{"data": "QmaozNR7DZHQK1ZcU9p7QdrshMvXqWK6gpu5rmrkPdT3L4"}],
     ).sign([alice.private_key])
 
-    app = App(b)
+    app = ApplicationLogic(b)
     app.init_chain(init_chain_request)
 
     begin_block = types.RequestBeginBlock()
@@ -365,7 +364,7 @@ def test_store_pre_commit_state_in_end_block(b, alice, init_chain_request):
 
     # simulate a chain migration and assert the height is shifted
     b.models.store_abci_chain(100, "new-chain")
-    app = App(b)
+    app = ApplicationLogic(b)
     app.begin_block(begin_block)
     app.deliver_tx(encode_tx_to_bytes(tx))
     app.end_block(types.RequestEndBlock(height=1))
@@ -470,39 +469,39 @@ def test_info_aborts_if_chain_is_not_synced(b):
     b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
-        App(b).info(types.RequestInfo())
+        ApplicationLogic(b).info(types.RequestInfo())
 
 
 def test_check_tx_aborts_if_chain_is_not_synced(b):
     b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
-        App(b).check_tx("some bytes")
+        ApplicationLogic(b).check_tx("some bytes")
 
 
 def test_begin_aborts_if_chain_is_not_synced(b):
     b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
-        App(b).info(types.RequestBeginBlock())
+        ApplicationLogic(b).info(types.RequestBeginBlock())
 
 
 def test_deliver_tx_aborts_if_chain_is_not_synced(b):
     b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
-        App(b).deliver_tx("some bytes")
+        ApplicationLogic(b).deliver_tx("some bytes")
 
 
 def test_end_block_aborts_if_chain_is_not_synced(b):
     b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
-        App(b).info(types.RequestEndBlock())
+        ApplicationLogic(b).info(types.RequestEndBlock())
 
 
 def test_commit_aborts_if_chain_is_not_synced(b):
     b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
-        App(b).commit()
+        ApplicationLogic(b).commit()
